Added review comments by Naresh
