+++ Requirement1/EmpAttendanceR1/EmpAttendanceR1.csproj
@@ -45,7 +45,9 @@
   <ItemGroup>
     <Compile Include="BaseEntity.cs" />
     <Compile Include="EmployeeEntity.cs" />
+    <Compile Include="EmployeePresentation.cs" />
     <Compile Include="GenericCRUD.cs" />
+    <Compile Include="IGenericCRUD.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
